+++ lib/src/core/navigation/app_router.dart
@@ -3,40 +3,38 @@ import 'package:nobook/src/core/navigation/app_routes.dart';
 import 'package:nobook/src/features/dashboard/view/screen/dashboard_screen.dart';
 import 'package:nobook/src/features/notes/view/screen/note_screen.dart';
 part 'app_route_paths.dart';
-class AppRouter{
+
+class AppRouter {
   static GoRouter get router => _router;
 
-
-
-  static void  goNamed (
-    String name,{
- Map<String, String> params = const <String, String>{},
+  static void goNamed(
+    String name, {
+    Map<String, String> params = const <String, String>{},
     Map<String, dynamic> queryParams = const <String, dynamic>{},
     Object? extra,
-    })=> router.goNamed(
-          name,
+  }) =>
+      router.goNamed(
+        name,
         params: params,
         queryParams: queryParams,
         extra: extra,
-    );
+      );
 }
 
-final  GoRouter _router = GoRouter(
-    urlPathStrategy: UrlPathStrategy.path,
-    routes: [
- GoRoute(
-      path: AppRoutePath.dashboard,
-      name: AppRoutes.dashboard,
-      builder: (context, state) {
-        return const DashboardScreen();
-      },
-    ),
- GoRoute(
-      path: AppRoutePath.note,
-      name: AppRoutes.note,
-      builder: (context, state) {
-        return const NoteScreen();
-      },
-    ),
-    ]
-);
+final GoRouter _router =
+    GoRouter(urlPathStrategy: UrlPathStrategy.path, routes: [
+  GoRoute(
+    path: AppRoutePath.dashboard,
+    name: AppRoutes.dashboard,
+    builder: (context, state) {
+      return const DashboardScreen();
+    },
+  ),
+  GoRoute(
+    path: AppRoutePath.note,
+    name: AppRoutes.note,
+    builder: (context, state) {
+      return const NoteScreen();
+    },
+  ),
+]);
