+++ lib/src/core/themes/color.dart
@@ -1,10 +1,13 @@
 import 'package:flutter/material.dart';
 
-class AppColors{
-   //#FFFFFF
+class AppColors {
+  //#FFFFFF
   static const Color white = Color(0xFFFFFFFF);
   //#999EAA
   static const Color grey100 = Color(0xFF999EAA);
   //#0C33FF
   static const Color blue = Color(0xFF0C33FF);
+
+  static const Color mCardColor = Color.fromRGBO(255, 255, 255, 1);
+  static const Color mBackgroundColor = Color.fromRGBO(251, 251, 251, 1);
 }
