+++ lib/src/features/notes/view/screen/note_screen.dart
@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
 
-
 class NoteScreen extends StatelessWidget {
   const NoteScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(backgroundColor: Colors.red,
-      body: Column(children: []),);
+    return Scaffold(
+      backgroundColor: Colors.red,
+      body: Column(children: const []),
+    );
   }
 }
